created config package to organize related files
