+++ src/main/java/config/PlayVoiceServiceConfig.java
@@ -1,4 +1,7 @@
-package view.services.playvoice;
+package config;
+
+import view.services.playvoice.IPlayVoiceService;
+import view.services.playvoice.PlayVoiceService;
 
 public class PlayVoiceServiceConfig {
     private PlayVoiceServiceConfig() {
