+++ src/main/java/usecase/acceptapplication/AcceptApplicationInteractor.java
@@ -1,6 +1,6 @@
 package usecase.acceptapplication;
 
-import dataaccess.DataAccessConfig;
+import config.DataAccessConfig;
 import dataaccess.IApplicationRepository;
 import dataaccess.IUserProjectsRepository;
 import dataaccess.IUserRepository;
