+++ src/main/java/usecase/createuser/CreateUserInteractor.java
@@ -1,6 +1,6 @@
 package usecase.createuser;
 
-import dataaccess.DataAccessConfig;
+import config.DataAccessConfig;
 import entities.User;
 import dataaccess.IUserRepository;
 import usecase.PasswordHasher;
