+++ src/main/java/usecase/editproject/EditProjectUseCaseFactory.java
@@ -2,7 +2,7 @@ package usecase.editproject;
 
 import api.embeddingapi.EmbeddingAPIInterface;
 import api.embeddingapi.OpenAPIDataEmbed;
-import dataaccess.DataAccessConfig;
+import config.DataAccessConfig;
 import dataaccess.IProjectRepository;
 import viewmodel.EditProjectPanelViewModel;
