+++ src/main/java/usecase/getloggedinuser/GetLoggedInUserInteractor.java
@@ -1,6 +1,6 @@
 package usecase.getloggedinuser;
 
-import dataaccess.DataAccessConfig;
+import config.DataAccessConfig;
 import dataaccess.ILoginUserDetails;
 import dataaccess.IUserRepository;
 import entities.User;
