+++ src/main/java/usecase/rejectapplication/RejectApplicationInteractor.java
@@ -1,6 +1,6 @@
 package usecase.rejectapplication;
 
-import dataaccess.DataAccessConfig;
+import config.DataAccessConfig;
 import dataaccess.IApplicationRepository;
 import dataaccess.IUserRepository;
