+++ src/main/java/usecase/searchforuser/LocalUserSearchObject.java
@@ -1,8 +1,7 @@
 package usecase.searchforuser;
 
-import dataaccess.DataAccessConfig;
+import config.DataAccessConfig;
 import dataaccess.IUserRepository;
-import entities.User;
 import entities.UserInterface;
 
 /**
