+++ src/main/java/usecase/searchprojectbyid/LocalProjectSearchByIdObject.java
@@ -1,6 +1,6 @@
 package usecase.searchprojectbyid;
 
-import dataaccess.DataAccessConfig;
+import config.DataAccessConfig;
 import dataaccess.IProjectRepository;
 import entities.Project;
 import entities.ProjectInterface;
