+++ src/main/java/view/CreateUserPanel.java
@@ -2,10 +2,10 @@ package view;
 
 import usecase.createuser.CreateUserController;
 import view.components.NumericTextField;
-import view.services.hovervoice.HoverVoiceServiceConfig;
+import config.HoverVoiceServiceConfig;
 import view.services.hovervoice.IHoverVoiceService;
 import view.services.playvoice.IPlayVoiceService;
-import view.services.playvoice.PlayVoiceServiceConfig;
+import config.PlayVoiceServiceConfig;
 import viewmodel.CreateUserPanelViewModel;
 
 import javax.swing.*;
