+++ src/main/java/view/MyProjectsPanel.java
@@ -7,10 +7,10 @@ import usecase.getprojects.ProjectData;
 import view.components.ButtonAction;
 import view.components.ButtonColumn;
 import view.services.SafeCastCollectionService;
-import view.services.hovervoice.HoverVoiceServiceConfig;
+import config.HoverVoiceServiceConfig;
 import view.services.hovervoice.IHoverVoiceService;
 import view.services.playvoice.IPlayVoiceService;
-import view.services.playvoice.PlayVoiceServiceConfig;
+import config.PlayVoiceServiceConfig;
 import viewmodel.EditProjectPanelViewModel;
 import viewmodel.MyProjectsPanelViewModel;
 import viewmodel.ViewManagerModel;
