+++ src/main/java/view/SwitchViewButtonPanel.java
@@ -1,10 +1,10 @@
 package view;
 
 import usecase.logout.LogoutController;
-import view.services.hovervoice.HoverVoiceServiceConfig;
+import config.HoverVoiceServiceConfig;
 import view.services.hovervoice.IHoverVoiceService;
 import view.services.playvoice.IPlayVoiceService;
-import view.services.playvoice.PlayVoiceServiceConfig;
+import config.PlayVoiceServiceConfig;
 import viewmodel.SwitchViewButtonPanelViewModel;
 import viewmodel.ViewManagerModel;
