+++ src/main/java/view/components/TagPanel.java
@@ -1,6 +1,6 @@
 package view.components;
 
-import view.services.hovervoice.HoverVoiceServiceConfig;
+import config.HoverVoiceServiceConfig;
 import view.services.hovervoice.IHoverVoiceService;
 
 import javax.swing.*;
