+++ src/test/java/usecase/edituser/EditUserUseCaseFactoryTest.java
@@ -1,15 +1,12 @@
 package usecase.edituser;
 
-import dataaccess.DataAccessConfig;
 import dataaccess.IUserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import viewmodel.EditProfileViewModel;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
 /**
  * Tests for the EditUserUseCaseFactory class.
